+++ src/Pages/HabitPage.js
@@ -1,13 +1,25 @@
-import { CalendarModal } from '../Exports';
+import { useState } from 'react';
+import { CalendarModal, Log } from '../Exports';
 import { HabitType } from '../consts';
 import PropTypes from 'prop-types';
 
-export default function HabitPage({ query }) {
-    if (query.type === HabitType) {
+export default function HabitPage({ type }) {
+    //const [ user, setUser ] = useState()
+    const [lastLog, setLastLog] = useState(' Never ');
+    const [streak, setStreak] = useState(0);
+
+    const handleLog = () => {
+        setStreak(streak + 1);
+        setLastLog(Date());
+        console.warn(lastLog);
+    };
+
+    if (type === HabitType) {
         return (
             <div>
                 <h1>Good Habit</h1>
                 <CalendarModal />
+                <Log streak={streak} onClick={handleLog} lastLog={lastLog} />
             </div>
         );
     } else {
@@ -15,13 +27,12 @@ export default function HabitPage({ query }) {
             <div>
                 <h1>Bad Habit</h1>
                 <CalendarModal />
+                <Log streak={streak} onClick={handleLog} lastLog={lastLog} />
             </div>
         );
     }
 }
 
 HabitPage.propTypes = {
-    query: {
-        type: PropTypes.string,
-    },
+    type: PropTypes.string,
 };
