+++ src/Pages/Home/Home.js
@@ -1,18 +1,8 @@
-import { useState } from 'react';
-import { Header, Log, NftBox, Intro } from '../../Exports';
+import { Header, NftBox, Intro } from '../../Exports';
 import './Home.css';
 import { HabitType } from '../../consts';
 
 const Home = () => {
-    //const [ user, setUser ] = useState()
-    const [lastLog, setLastLog] = useState(' Never ');
-    const [streak, setStreak] = useState(0);
-
-    const handleLog = () => {
-        setStreak(streak + 1);
-        setLastLog(Date());
-    };
-
     return (
         <div>
             <Header title={'HabiPets'} />
@@ -21,7 +11,6 @@ const Home = () => {
                 <NftBox type={HabitType.GOOD} />
                 <NftBox type={HabitType.BAD} />
             </div>
-            <Log streak={streak} onClick={handleLog} lastLog={lastLog} />
             <button onClick={() => console.log(process.env)}></button>
         </div>
     );
